// src/types/CsvValue.ts
export default interface CsvValue {
  [key: string]: string;
  EnterpriseFlag: string;
  RecordNumber: string;
  MSACode: string;
  CensusTract: string;
  TractIncomeRatio: string;
  BorrowerIncomeRatio: string;
  LTV: string;
  PurposeOfLoan: string;
  FederalGuarantee: string;
  BorrowerRace: string;
  CoBorrowerRace: string;
  BorrowerGender: string;
  CoBorrowerGender: string;
  NumberOfUnits: string;
  UnitAffordabilityCategory: string;
}